Wger
# Wget Exercise 2

I successfully downloaded the first page activehistory to my new directory, but when it comes to downloading every paper, I ran into trouble. I inserted the recursive command and the active history URL into the wget **[options] [URL]** formula, but ended up with an "invalid IPv6 numeric address". However, when I read further, and was made to understand the double-dash command, the input to change depth of pages, and the download speed, I finished with the full formula and had a successful (but lengthy) download with: 
> wget -r --no-parent -w 2 --limit-rate=20k http://activehistory.ca/papers/