twarc work
When trying to install Twarc to work with the DH Box, I received the message that it was “cleaning up” and the command python setup.py egg_info failed with error code 1, and that it was storing debug log for failure in /home/sherlynassam/.pip/pip.log. Unsure what this meant, I tried to make a sudo command out of it, but I received the same message. 

I went to Zulip to see if my classmates had run into any problems. I followed the link Danny Jooste supplied to Lauren Rollit, then tried to fix my setup tools, according to a comment, by typing “pip install –upgrade setup tools” but I got the messaged that I can’t rollback setup tools because there were not uninstalled. After that, I added “sudo” in front of it, and they were successfully installed.