twarc
I tried to move on to step 12 with $ twarc configure, but it was not recognized as a command. I tried to download Twarc again, by typing the command in step 11, and then I was prompted to provide a consumer key, consumer secret, access token, access token secret. 

I was expecting there to be a lengthy list of downloads, but it stopped after half a dozen lines, reading: “AttributeError: ‘module’ object has no attribute ‘ReadTimeout’. Either way, I proceeded to the next step that was limited to hist3814o instead of canada150, and got the same attribute error message.